added match and matchAny to junify module
